# backend/requirement/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: AI Task Dashboard API
  description: >
    Automatically derived CRUD API documentation based on `Data requirements.md`.
    All responses follow the standardized API response envelope.
  version: 1.1.0
servers:
  - url: /api/v1
    description: Main API Server
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get all notifications for user
      description: Notification stream requiring strict pagination with limits and offsets.
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of notifications including contextual derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items:
                              $ref: '#/components/schemas/NotificationResponse'
    post:
      tags: [Notifications]
      summary: Create a notification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationBase'
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notification:
                            $ref: '#/components/schemas/NotificationResponse'
...
  /notifications/{id}:
    get:
      tags: [Notifications]
      summary: Get notification by MongoDB _id
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '200':
          description: Notification object with derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notification:
                            $ref: '#/components/schemas/NotificationResponse'
    put:
      tags: [Notifications]
      summary: Update a notification (e.g., mark as read/isSeen)
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationBase'
      responses:
        '200':
          description: Notification updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notification:
                            $ref: '#/components/schemas/NotificationResponse'
    patch:
      tags: [Notifications]
      summary: Update notification status (isSeen)
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [isSeen]
              properties:
                isSeen:
                  type: boolean
                  description: Mark notification as seen or unseen
      responses:
        '200':
          description: Notification status updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notification:
                            $ref: '#/components/schemas/NotificationResponse'
    delete:
      tags: [Notifications]
      summary: Delete a notification
      parameters:
        - $ref: '#/components/parameters/MongoIdParam'
      responses:
        '204':
          description: Notification deleted
  /notifications/user/{userId}:
    get:
      tags: [Notifications]
      summary: Get notifications by user ID
      operationId: getNotificationsByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID (recipient)
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of notifications for the specified user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items:
                              $ref: '#/components/schemas/NotificationResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /notifications/user/{userId}/count:
    get:
      tags: [Notifications]
      summary: Get notification count for a user
      operationId: getNotificationCountByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: 24-character hexadecimal MongoDB User ID (recipient)
        - name: isSeen
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by seen status. If omitted, returns count of all notifications.
      responses:
        '200':
          description: Notification count (filtered or total)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
components:
  parameters:
    MongoIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 24-character hexadecimal MongoDB Object ID
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: Amount of records to return for pagination
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
      description: Amount of records to skip for pagination
  schemas:
...
    ApiResponseEnvelope:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: string
          enum: [success, failed]
          description: Indicates whether the request was successful or failed
        message:
          type: string
          description: Human-readable success or failure message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 date-time of the response
        data:
          type: object
          nullable: true
          description: Domain-specific payload. May be null on failure.
    CountApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                count:
                  type: integer
                  description: The total count

    # ------------------
    # MINIMAL SCHEMAS (FOR REUSE)
    # ------------------
    UserMinimal:
      type: object
      required: [_id, name]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        name:
          type: string
        email:
          type: string
        companyTechRole:
          type: string
        role:
          type: string
          enum: ['ADMIN', 'MEMBER', 'USER']
    ChallengeMinimal:
      type: object
      required: [_id, title]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        virtualId:
          type: string
          description: CH-XXX format
        title:
          type: string
    IdeaMinimal:
      type: object
      required: [_id, title]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        ideaId:
          type: string
          description: ID-XXXX format
        title:
          type: string
...
    NotificationBase:
      type: object
      properties:
        type:
          type: string
          enum:
            - challenge_created
            - challenge_status_update
            - challenge_edited
            - idea_edited
            - challenge_upvoted
            - idea_upvoted
            - challenge_commented
            - idea_commented
            - challenge_subscribed
            - idea_subscribed
            - challenge_deleted
            - idea_deleted
        fk_id:
          type: string
          nullable: true
        userId:
          type: string
        initiatorId:
          type: string
          description: Mongo hex ID of the user who triggered the notification
        isSeen:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time

    # ------------------
    # RESPONSE SCHEMAS combining Base + Derived Fields
    # ------------------
...
    NotificationResponse:
      allOf:
        - $ref: '#/components/schemas/NotificationBase'
        - type: object
          properties:
            _id:
              type: string
            description:
              type: string
              description: "(DERIVED) Dynamically generated based on Type and initiator"
            linkedEntityDetails:
              type: object
              oneOf:
                - $ref: '#/components/schemas/ChallengeMinimal'
                - $ref: '#/components/schemas/IdeaMinimal'
              description: "(DERIVED) Details of Challenge or Idea linked via fk_id"
            recipientDetails:
              $ref: '#/components/schemas/UserMinimal'
            initiatorDetails:
              $ref: '#/components/schemas/UserMinimal'
...